+++ odoo/songs/install/data_full.py
@@ -125,6 +125,12 @@ def import_bank(ctx):
 
 
 @anthem.log
+def import_bank_account(ctx):
+    """ Importing bank account partners from csv """
+    load_csv(ctx, 'data/install/bank_account.csv', 'res.partner.bank')
+
+
+@anthem.log
 def main(ctx):
     """ Loading full data """
     import_users(ctx)
@@ -142,4 +148,5 @@ def main(ctx):
     import_workcenter(ctx)
     import_rma_cause(ctx)
     import_bank(ctx)
+    import_bank_account(ctx)
     return